Start a branch w/ preliminary shot at a better fix for bug 2102. These changes are also desired to aid addition of local firehose.
I have done a preliminary version of the changes to move the payload sends
of AMReplyLong from a dynamically allocated EQ to the gasnetc_SAFE_EQ.

This change eliminates PtlEQ{Alloc,Free} from the critical path and therefore
removes the mutex Dan added (comment #5) as a work around (since the only
remaining calls are in the init and exit paths).  It also allowed use of the
RARSRC MD (eliminating dynamic MD allocation from the critical path) for
in-segment source addresses.

So for the best case of an in-segment source, we are eliminating from the
critical path a gasnet-level mutex, an EQ alloc-free pair and an MD bind-unlink
pair.  Since our "testam" benchmark uses only in-segment sources for
AMReplyLong, these changes results in significant improvements in reported
bandwidth of an AM Request-Reply flood test (measurably higher peak, and nearly
doubled bandwidth in the important middle range from 4k to 16k).

So far testing on a system w/ PE 2.0.24b with the gasnet PAR tests show no
correctness problems in a DBG build.  Additional testing (including checks for
performance regression in an OPT build) will be done before this code is merged
to the CVS HEAD.