enable a "forced" dcmf build
sicne mondays are the weekly maintence days of the BG/P system at ANL i've set up a mode where we can at least compile the code (and not link) assuming we havea  local copy of the header files. This will at least allow development to continue on days where machine access is down. For the final release this code should be taken out. This only intended for dev purposes.
CVSMAIL: CC: