testam, testsmall, and testbarrier pass... still problems w/ testlarge
looks like testam, testsmall, and testbarrier pass in both debug and opt builds However there's still a problem w/ testlarge. The main bug that we hit was that the DCMF messaging layer and GASNet were being compiled w/ different compilers and screwing up the alignment. It would be nice if there was a check during configure time that this didn't happen.

TODO for AMS:
+ Pass all gasnet tests!
+ AMMediums still do a "malloc" for bounce buffers and we need to switch that toa  freelist (and possibly flowcontrol to prevent unbounded growth). I think the important question to ask is if DCMF has flow control underneath the convers it uses for the send operation to prevent some other resource exhaustion problem. f thats the case i think we can safely just deal w/ a free list.
+ Performance tuning

CVSMAIL: CC: