fixed gasnetc_exit to allow for non-collective exit
BG/P sends a SIGTERM to the other nodes when a exit(1) is called on the local node. Thus we trun all non-zero exit codes to exit(1) so we can get this behavior. I"ve sent out mail asking if there is a better solution but until then this seems like the way to go. all testexit tests pass except for the ones that return from main rather than call exit


CVSMAIL: CC: