overhaul of how the tuning process is handled
the loops that iterate over the paramter space are now within GASNet rather than in the user level.This enables us to better control how the parameter space is searched over through environment flags in the future rather than relying on exposing awkward paramter space iterators to the user.

The testcolltuner routine just calls this tuning routine which returns the best algorithm and paramters as opaque integer things taht we will pass back to GASNet when the job is loaded up.

Todo still is to make the decision trees work
CVSMAIL: CC: