I've started "BUG651_BRANCH" in CVS to work on the code common to this issue and the SysV supernode discovery.
The current idea is an array of gasneti_nodes elements of type gasnet_node_t.  Each element nodemap[i] gives the lowest gasnet node number to share a physical node with node i.  For instance, with 5 gasnet nodes started 2-per physical node:
   nodemap[] = {0, 0, 2, 2, 4}
Note that I've chosen to implement this in terms of gasnet_node_t regardless of the type of identifier used to derive it: hostid, MAC addr, etc.

My expectation is that this array can be generated once in/from gasnetc_init() and passed to the mmap probe and to the SysV initialization code.  The array would then be free()ed no later than the end of gasnetc_attach().

I have implemented a generic
  extern void gasneti_nodemap(gasnet_node_t *nodemap,
                              gasneti_bootstrapExchangefn_t exchangefn);
with provisions for conduit-specific and platform-specific overrides.

The generic gasneti_nodemap() is implemented in terms of
  extern gasneti_hostid_t gasneti_gethostid();
This also has a generic implementation for systems which pass a configure-time probe for gethostid().  Again, conduit-specifc and platform-specific overrides are possible.

Note that if a conduit- or platform-specific override if provided for gasneti_nodemap(), then gasneti_gethostid() is not required.  This can be changed if we find need for the ids elsewhere in gasnet.

My next step is to re-work the vapi-conduit code that does the equivalent work to construct a "conforming" nodemap array.  Since vapi-conduit already does an exchange of address info for wire-up, there is no need in vapi-conduit to generate additional gasneti_hostid_t, nor to perform additional communication to propagate it.