changed the code such that only thread 0 polls and initiates the collectives
this had SIGNIFICANT performance improvements on a modest 4 node run on hive w/ two pthreads per node. We have dramatically decreased the lock contention and other such factors that were detracting to collective performance in PAR builds. There are a couple more ideas in the works such as better management of the handles but this is a start.
CVSMAIL: CC: