Start a branch for work on Bug 2461 - PTL_ME_IN_USE runtime crash in portals/par
This commit is the first working attempt at fixing the problems described in
bug 2461.  It does not yet deal with the CB errors in Bug 2461.

The basic approach in this is to
1) Ensure ref count on ReqRB buffers is incremented atomically with respect
   to the poll of the AM_EQ.  The ensures that we cannot ever miss the
   increment from an event prior to the current one.  Doing so would have
   resulted in refreshing the buffer while an AM in the buffer was still
   "live".  (This addresses Dan's Comment #2 in the bug report)
2) Switch on AUTO_UNLINK to ensure we don't try to refresh the buffer while
   there are still PUT_END events (corresponding to AM requests) in the EQ.
   Doing so would have, again, resulted in refeshing the buffer while there
   were still AMs using it.
3) Use of AUTO_UNLINK also ensures that the buffer is refreshed exactly once.
   As observed in comments #6 and #7, the multiple unlink attempts made in
   the Manual unlink case were leading to multiple MDs describing the same
   memory.
4) I have also factored the bufptr lookup code (for performance).

This version still contains debugging output, disabled via "//" comments.
These will be convereted to TRACE/STATS calls and the "//" removed before
this can be merged to the HEAD.  Additional cleanups are planned.


Unfortunately, Bug 1976 (testvis fails when GASNET_VIS_AMPIPE set) appears
unaffected by these changes.