fix subordination in multiaddr collectives
when a multi address collective calls a single address collective we incorrectly reforward the logic to another multiaddr collective causing a deadlock. the fix is to just have the single addr collective only forward to the multi addr one if it's not a subordinate (i.e. there are other theads that will post the address)
CVSMAIL: CC: