Use PtlMDUpdate in the firehose move callback when we have both unpins and pins.
When I initially did local firehose support for portals-conduit, I found
that PtlMDUpdate() would always return PTL_INVALID_EQ, even though the
documentation from Sandia says that PTL_EQ_NONE was a valid argument.
At the time I just noted the problem in a comment and went on w/o worrying
about the loss of this potential optimization.

I am now creating a one-element "EMPTY_EQ" for use in place of PTL_EQ_NONE
as a work-around for the PTL_EQ_NONE problem.  The result is that time for
a move callback which performs an unpin and a pin is reduced by 5% (from
7.73 +/- 0.03us to 7.36 +/- 0.02us).

The return of PTL_INVALID_EQ when passing PTL_EQ_NONE has been submitted
to NERSC (assigned Incident #090512-000009) for reporting to Cray through
normal support channels.