Fix the "divide by N-squared" bug discovered by Filip.
Filip has observed that even after my mmap probe had determined the total
amount of memory that could be mmap()ed by multiple processes on a node,
and had divided that by the number of processes, he sometimes needed to
divide AGAIN by the number of processes to get sucessfull mmap()s.

I have determined that my mmap probe was doing exactly what I thought it
should, but also determined that for the SYSV code that something MORE
was needed.  The origin of Filip's problem is not in the limit on how
much memory can be pinned (which I had computed correctly), but rather
in fitting the mapping of all of that memory into EACH process (as opposed
to 1/N of it in each process).  This was particularly problematic on
32-bit platforms where you might have 2- or 4-GB of phsical memory (or
more), but no hope of mapping that much into the 32-bit address space.

This commit adds additional logic to further probe the limit on how much
can be mapped into a given address space N times (N = processes per node).