Significantly reduce chance of leaking shared memory.
As noted in a previous commit, we had the possibility of leaking the
shared memory if there was an exit between the start of gasnet_init()
and the end of gasnet_attach(), incuding both errors and the possibility
of the client code exiting (e.g. due to insufficient segment size).
There was also a potentially significant amount of counduit-specific code
that could run while the shared memory was "vulnerable" to leaks, and
modifying all such code to cleanup before gasneti_fatalerror() and/or
failed assertions is undesirable/impractical.  [I also dismissed, for
now at least, the idea of forcing gasneti_fatalerror() and
gasneti_killmyprocess() to perform shared memory cleanup.]

This commit reduces the "window of vulnerability" to the following four
conduit-independent functions which manipulate mmap()s:
  gasnetc_init_sysv()
  gasneti_mmapLimit()
  gasneti_segmentInit()
  gasneti_segmentAttach()
Upon return from each of these, there are no POSIX shared memory objects
which have not been shm_unlink()ed.

Each of these functions takes one or more conduit-specific bootstrap
collectives.  So, there is still some small amount of conduit-specific
code that might leak shared memory in the event of a gasneti_fatalerror().
Additionally, these four functions have not yet been fully audited to
ensure that they shm_unlink() before a gasneti_fatalerror().  However,
the amount of code that requires audit it now MUCH smaller.

Note that I can run "make testexit-par run-testexit" and while there
are still about 10 tests that hang, there are ZERO shared memory files
leaked (and just 2 zero-byte files in /tmp).  Before this commit the
two tests that exit between init and attach were leaking not only the
zero-byte file, but also the shared memory files.

NOTE: The SYSV code does not yet work under Darwin, which appears to
disallow resizing of a shared memory object.  Making things work for
Darwin will require a shm_unlink() before each "resize" in order to
get a "fresh" object. This commit moves us in that direction.