added field to algorithm index
we now support a n_requirements field which contains a list of flags that a collective will not support (ie. GASNETE_COLL_THREAD_LOCAL). It would have takena  lot of duplicate engineering effort to get THREAD_LOCAL to work for smp collectives so we punt on it w/ this hack for now
CVSMAIL: CC: