CRASH RECOVERY REPLAY OF... Paul H. Hargrove <PHHargrove@lbl.gov>  committed changes to cvs project: gasnet (branch InterNode_SYSV_branch)  on Thu Sep 10 05:08:46 PDT 2009
 --- Log message ---

Avoid a possible race (not yet observed in practice) between 2 users of "vnet".
The "early" System-category AM in vapi-conduit wants to use the "vnet" for
AMs, but the gasneti_mmapLimit() call that follows contains a call to
gasneti_pshmnet_bootstrapBroadcast(), which uses the "vnet" in a "raw" mode
without any AM headers.
I suspect that this didn't cause any problems yet only because we were
lucky enough on timing to always have polled the AM off the queue before
the "raw" use in mmapLimit().