Bug 2809 (ibv-conduit SRQ support) "Step 2" This commit completes part 2 of the SRQ development plan in bug 2809:   Step 2) keep linear memory usage for low node count and flatten   out at some max number of buffers (env var). The env var is GASNET_RBUF_COUNT to match naming of other vars, and is documented in README. This commit also includes a portion Step 3.3 by logging the actual buffer count and the number w/ and w/o SRQ.
Testing confirms that I can run MOST of the GASNet tests w/ 4 nodes and
just 2 AM recv buffers (fewer than the number of peers).  However, the
lack of flow control does allow this (unrealistic) reduced resource case
to deadlock testinternal and testcoll.  In the case of testinternal the
deadlock has been traced to the progress functions test and the way in
which it mixes RDMA with AMs.  Because AMs are "stuck" waiting for recv
resources to become available on the target node, a blocking RDMA op
becomes blocked waiting for send resources to become available on the
initiator - which will not happen.  I did not attempt to drill down to
the cause of the testcoll deadlock as the code is quite a bit more
involved.  However, I suspect the problem is the same - blocked AM ops
"own" all the send resources.

The deadlocks described above are EXPECTED/ALLOWED in the first (June
30, 2010) deliverable.  The purpose of the 2nd (Sep 30, 2010) deliverable
is to select an appropriate flow-control (or traffic separation) mechanism
to resolve this.  ALSO, the deadlock is nearly impossible to reproduce for
any realistic number of AM recv buffers.