Make a distinction between the number of QPs per peer, and the number used "normally".
This is in preparation of the flow-control work which is expected to add one or
more QPs per peer that are dedicated only to AM Request traffic (as well as the
possibility that AFTER the QP array is allocated we might determine that SRQ
will NOT be used and thus we'll have unused spaces in the table).

This adds nothing to the length of the critical paths (where we've just replaced
one variable by a different one) and only mildly complicates the initialization
code which must now skip over any unused slots.  I'll need to rework some of
that init code to setup two distinct "classes" of QP for the SRQ flow control, but
have not progressed far enough in the design to start those modifications yet.
There will be resource calculation changes as well, since the hca struct fields
that count qps have not be modified to make any normal-vs-other distinction yet
(they just count "non-NULL" at the moment).

Note that this was tested for both IBV and VAPI by manually setting
   gasnetc_num_qps = gasnetc_normal_qps + 1
early in the init code, and running all the standard GASNet tests.