Next step on the road to SRQ w/ flow-control
This commit creates a second SRQ to separate the pools of recv buffers used
for incomming Requests and Replies.  Along with placing them on separate QPs
this is the second major change required to eliminate deadlock, by ensuring
that Reply traffic can always make progress independent of Request traffic
(which is permitted to block awaiting resources).
This commit also adjusts the setting of the SRQ depth and the corresponding
allocation of recv buffers to be "tight" (no waste).  Additional work is
still needed to make sure that the size of the Send Work Queues and the
Completion Queues is correct (I know the Send WQ is too deep and suspect
the corresponding CQ is too).
Note that there is still no accounting of AM traffic at the initiator - it
is still unclear to me if that is required to avoid deadlock, or just desired
to limit conjestion.  Either way it is expected eventually.