Yet another step toward SRQ+F/C
This commit moves the completion records for received Replies to the "snd_cq".
This ensures that we poll Replies independently from Requests, and may lead
to better progress in some applications.  This is similar to the "SAFE_EQ"
in portals-conduit.  This is imortant because the Reply headers may include
credits (though not yet) needed to unblock.
This commit also ensures the CQ depths are set correctly both w/ and w/o SRQ.