in gasneti_pshm_init(), testing if the number of PSHM nodes is always smaller than GASNETI_PSHM_MAX_NODES.
In case when exchangefn is not NULL, we will be fine since it will act as a barrier, but
a problem might occurwhen exchangefn is NULL: one (or more) supernodes could fail here, but
the rest of them will continue allocating shared memory and fail upon allocating the shared
segments (due to gasneti_fatalerror() call). This might leave some shared segments in the
system. We could put a global barrier that would stop the progress until we know all
supernodes are good. But that would affect the initialization performance.

CVSMAIL: CC: