Correction to the code I added last night for the known race condition.
I missed the possibily that the ACK had been received earlier in the
same Poll and therefore was not yet acted upon.  This fixes that by
taking a different approch to determining when we are not the first thread
to notice the implied ACK.  I have also made that logic conditional on
GASNETI_THREADS, since it CANNOT happen in a single-threaded build (and
I've added an assertion to check that).