Remove various unnecessary complications from the static connect code.
This commit removes the connection randomization and the gratuitously
split loops that were used to help "model" the behavior of the dynamic
connect code before it was written.  Now that the dynamic code is complete
these things are not needed and just slow the code and would make future
maintenance more difficult.