Merge MIPS-specific fixes from HEAD to the BUPC-2.14 release branch. This commit applies the 3 changes listed at the end of this message.
These are being applied to the release branch at this very late date because
the problems they fix represent regressions introduced with the addition
of PSHM support in GASNet 1.14 (Nov 2009).  This regression went unnoticed
for two full years because, as described below, we don't normally test much
in the way of MIPS systems other than SiCortex.  However, the effected
system(s) did work prior to addition of PSHM support (with the exception of
the "o32" ABI which was always broken prior to this commit).

I have only considered merging these changes because they are all conditional
on MIPS, and thus cannot break other platforms (short of mistakes like
introducing unmatched #if's).  So, in order to "qualify" this for inclusion
in the release I have tested on:

System #1: SiCortex at ANL (MIPS64EL/Linux)
  gcc-4.2.3    ABI=64
  pathcc-3.3a  ABI=64
System #2: Octane at Intrepid Tech. (MIPS64/IRIX)
  gcc-3.3.0    ABI=n32,64
  gcc-4.2.4    ABI=o32,64
System #3: Malta (MIPS64/Linux)
  gcc-4.3.2    ABI=o32,n32,64

The remaining compiler+ABI combinations are not available because the necessary
compiler support and/or runtime libraries are not installed/available.

========
MIPS-specific fix 1 of 2
The correctness of our 64-bit atomics on MIPS is a functions of the ABI,
rather than of the ISA.  This commit corrects the atomics code to avoid
incorrect atomics when building for the "o32" ABI on a 64-bit MIPS CPU.
This has been broken for a long time now, but we don't normally test any
platforms that would have shown the problem.
========
Cosmetic fix to my previous commit
========
MIPS-specific fix 2 of 2
This commit fixes configure-time detection of non-native atomics on MIPS,
which is performed as a prerequisite for enabling PSHM support.
The availability of native atomics on MIPS depends on the ISA level (need
the ll/sc instructions), which can be determined from the value of the
pre-processor token _MIPS_ISA.  However, the "normal" behavior of most (all?)
compilers is to
  #define _MIPS_ISA _MIPS_ISA_MIPSn
for some integer 'n'.  These _MIPS_ISA_MIPS* constants are then defined in
the header sgidefs.h.  However, since this header was not included as part
of the configure-time probe for non-native atomics we were ALWAYS concluding
that we had non-native atomics.  This commit moves the probe for the header
sgidefs.h earlier in configure, and relocates the corresponding #include from
gasnet_basic.h to gasnet_asm.h (which is actually more appropriate anyway).
This problem went unnoticed until now because neither of the MIPS platforms
normally tested (a SiCortex and an SGI Octane running IRIX) support POSIX
shared memory, and thus the probe for non-native atomics never ran on MIPS
prior to my recent testing on a big-endian MIPS64/Linux platform.
========