Start a branch to explore potential benefit of atomic_swap
This is now a key operation in the AMPSHM code (to enqueue) and might also be
used (after some other non-trivial changes) in UPCR's MCS-locks implementation.
The implementation is incomplete, but has a "native" version for x86/x86-64
and the "default" version in terms of CAS for other CPUs.

As noted in the comments, the initial results don't show any benefit.