"Full" immplementation of "RDMADISSEM" barrier
This uses blocking 16-byte puts, but a non-blocking version could be
built w/ sufficient "machinery" to manage the handles.
The next step is to do some performance profiling to see if/where the
new implementation represents an improvement, and to determine if the
additional work of a non-blocking implementation is worth pursuing.