Fix RDMA-dissemination barrier NB-Put "breakage" with PSHM
Code which added non-blocking Puts was incorrect with respect to the PSHM
heirarchical barrier interaction.  In particular, it incorrectly assumed
that ONLY the supernode-representative ever performs Puts.  However, the
reality is that the last node the arrive at the barrier will always be
the one to issue the Put for step==0, and that can be ANY node.