Add "DISSEM" barrier (pseudo-)algorithm. It selects RDMADISSEM when Put is native, or AMDISSEM when using the ref-extended implementation of Puts.
Measurements on several conduits with native RDMA (gemini, pami, ibv) show
RDMADISSEM to be measureably better when the node count is 4 or more, but
with 2 or 3 nodes the results are mixed.  Meanwhile, on MPI conduit (using
the ref-extended for Puts) the AMDISSEM is always the winner due to the
lower overheads.

It may be posible to make this DISSEM barrier logic always select AMDISSEM
in a 2-node run, but I think AMCENTRAL is actually the better choice there.