Remove memory fences which have been unnecessary since 2006!
When progress functions (the barrier "kick" functions) were added, it became
necessary to add HSL's in the functions which previously required fences to
correctly order certain reads and writes.  Since that time the fences which
seprated access to amdbarrier_recv_value_present and amdbarrier_recv_value
have been unnecessary because they all take place inside of HSL critical
sections.

Additionally, the WMB prior to writing to amdbarrier_step_done in the AM
handler can be eliminated by simply moving that write to AFTER the unlock
(which must include a WMB).  Such a transformation is safe, since the
uniqueness of the (phase,step) ensures the write cannot conflict with
any other.