Move source of Puts in RDMADISSEM barrier from the stack to the aux segment.
This uses the padding that exists to keep active inboxes on distinct cache
lines, and so comes "for free".  By using the padding in the inactive
phase we also avoid creating contention with the incomming traffic.

By using the aux segment not only do we now have an in-segment source (to
eliminate any need for dynamic pinning of a stack page), but we also now
have buffers with sufficient lifetime to use bulk operations in the
non-blocking case.  However, most conduits already have optimizations for
small non-bulk puts (inline or bounce-buffer) which minimize the actual
advantage of both new properties (in-segment source and bulk xfer).

The use of distinct buffers for each step will (in a later commit) allow
for potential communication/communication overlap.  So, this is a move
in the right direction even if the inline and bounce-buffer optimizations
eliminate the other advantages.