Implement communication-communication overlap in RDMA-DISSEM barrier.
We now always use the non-blocking bulk Put to send out notifications.
In the multi-threaded case (including SEQ w/ conduit threads) we MUST
sync the handles before leaving the initiator (either _notify or _kick).
The single-threaded case is essentially unchanged.

The difference is that now the multi-threaded case can (and often will)
issue multiple non-blocking Puts from the kick function before it does
any syncs, thus producing communication-communication overlap which was
not present before.  The advantages of this don't show on ibv-conduit
(which has a conduit thread) with testbarrier, where the results are
essentially unchanged.  However, with testbarrierlate (where the
bottleneck is in the kick by the node which arrives last at notify or
wait) the improvement is significant even with as few as 5 nodes (the
least that requires issuing more than one Put from the kick function).