Increase RDMADISSEM comm/comm overlap in Wait for multi-threaded case.
For the case of a multi-threaded client (including SEQ + client threads)
the RDMADISSEM code previously synced all outstanding handles at the end
of each call to the "kick" function.  This was necessary because in general
we couldn't know that the calling thread would ever call the barrier code
again.  However, in the common case of a call to gasnet_barrier_wait() we
DO know that the same thread will continue to call the kick function until
the barrier has been completed.  So, this commit leverages that knowledge
to move the sync of handles created by a multi-threaded build's call to
Wait to take place at the end of Wait, just as in a single-threaded build.

Another commit (soon now, I hope) will do the same for the lock acquisition
used to ensure correct operation in the presence of concurrent "kickers".