REfix the problem addressed by the previous commit.
The problem of "slow" threads holding on to slots in the handle array is
now resolved by using automatic variables instead of the per-team array.
This allows the reduction of the critical section back to the point where
it excludes the Puts and their syncs.

The re-rentrance protection has been removed.  This means that when a Try
or Wait (in a multi-threaded build) calls the Kick function the call to
wait_syncnb_all() will invoke the Kick function.  It is not clear if this
is DESIRABLE, but at least with the reduced critical section it is no
longer ERRONEOUS.  This re-entrance protection might be restored later,
but not until I've studied the peformance implications.