Disable AMDISSEM and AMCENTRAL barrier progress functions "early" in Wait
Testing shows that by turning off the progress function before the Wait
caller begins spin polling reduces barrier latency in testbarrier-par
regardless of whether polling threads are used or not.  This is presumably
due to two things:
1) Reduced contention from the polling threads
2) The BLOCKUNTIL macro includes an AMPoll which would run the progress
   function IMplicitly in addition to the EXplict call in its argument.
   This double-call may have increased the latency between the actual
   completion and the return of control to Wait.

Notably the AMRDMA barrier appears to get WORSE performance when the
progress function is disabled early in this manner, OR when it is just
prevented from entering the Kick concurrent with a Try or Wait.