Begin "GASNET_ARIES_branch" This commit contains the minimal changes required to get gemini-conduit running at all on Aries, but is KNOWN TO STLL BE INCORRECT.
The issue is that GNU_SmsgSend no supports the mode in which the header
and payload are sent synchronously.  So, we *must* use AUTO_RETRANSMIT
mode.  That requires that we cannot overwrite (or free/unmap) the header
or payload until after we process the event from the Cq.  We are currently
TOTALY ignnoring that requirement and therefore might send total garbage
in an AM.