Start a branch with an alternative implementation of AMs
This version implements the AMs directly in terms of FMAs.
The "key" is the "PUT_W_SYNCFLAG" which is a sort of signalling put
that generates CQ entry on target and writes a single-word flag
there to mark "full" mailboxes (after they are fully written), all
in a single operation together with the xfer of the full payload.

This "re-write" uses less registered memory, and has smaller
source code and object code.

Current benchmarks don't favor use of this implementation on Gemini
(where we must now copy AM payload to make contiguous w/ header):
    upto 10% BETTER 0-arg/0-byte round-trip la
    upto 10% WORSE round-trip flood b/w
    upto 40% WORSE one-way flood b/w

But things look favorable on Aries (where have always copied AM
payload for local completion):
    upto 14% BETTER 0-arg/0-byte round-trip latency
    upto 5% BETTER round-trip flood b/w
    upto 6% BETTER one-way flood b/w

Note that on Aries we have a reproducable "issue" with how we infer
the available credits for Smsg sends, which I've never been able to
reproduce on Gemini.  So, this may become the implementation for
Aries while keeping the old one for Gemini.  However, this branch
holds only the NEW implementation, having entirely removed the one
based on GNI_SmsgSend().

Some additional tuning and study is in order before attempting a
dual-mode version.