This is hopefully a branch commit for moving the AM buffers into a smaller pool of buffers which are held until the reply. As a result, the buffer allocation policies for the various paths.
Its expected that there will be some refactoring before merging
to head to minimize the patch footprint and for efficiency.
In addition there are some suspect uses of slot constants,
and potentially incorrect timing around the per-peer enqueue
and dequeue of buffers.

Overall the changes as they stand seem to have a 10% performance
penalty for small messages on both edison and hopper, presumably due
to the list manipulations for the posted buffer.

This patch doesn't include the actual change to the smsg pool.

CVSMAIL: CC: