Remove logic for delayed return of credits.
The motivation behind delayed return of credts was to reduce the number
of implict reply messages on the wire for certain traffic patterns.
However, the new buffer management in this branch requires that the
implicit reply always be sent.  Thus the delayed return of credits no
longer serves any purpose (and was just complicating the code).