First pass merging Eric H's cleanups
This commit narrows the type for a slot number to 16 bits, adds proper
casts to the 2 constants of this type, and bounds-checks the env var
agains the range of this type.

In contrast to Eric's patch, this commit silently reduces the number
of AM buffers to the range of the type, rather than generating a fatal
error when the value is too large.