Reduce AM buffer memory by approximately factor of two.
This commit is one of the motivations for this branch: to reduce by
a factor of 2 the prefactor for the O(N) portion of the memory used
for Active Messages.  Given a network depth of 'd', we longer need to
have enough memory reserved to receive both 'd' Requests and 'd'
Replies from each peer.  Instead we need space for only the 'd'
Requests per peer, and a fixed (independent of number of peers) pool
of buffers for Replies.

This commit just reduces the allocation.  The code on this branch has,
for a while now, only needed the smaller buffer size.