Restructure header to restore pre-branch length of Short and Long AMs
With the elimination of "is_req" and "credit" fields from the header, we
can devote a distinct "command" value to packed-long messages and thus no
longer need to use "misc" to identify them.  That leaves only the nbytes
of a Medium as a use of "misc".  So, by moving that to the Medium-specific
part of the message, we can repurpose those 16 bits for "reply_slot" and
restore the GC_Header to the same 32-bits as was used prior to the start
of this branch.  This means that only Medium messages are potentially any
larger (by 4-bytes) on this branch than in HEAD.