Committing Eric Hoffman's "Notify Word" work to "ambuffer" branch.
This patch modifies the AM buffer allocation to use a single shared pool
for the initiator, while preserving per-peer buffers for the target. The
goal is to reduce the amount of memory reserved for active messaging on
a per-node basis.

In addition it moves the synch notify field into a separate per-peer ring
buffer from the packet head. This 'notify' word contains the allocations
for that messsage by slot on both the initiator and the target. This allows
AM requests and replies to be serviced from the same polling process.

The initiator's buffer is used to format the request, and will contain the
reply. The target buffer is initially filled by the request and is used
to format the reply.

In addition, the empty credit message is sent using a fma put of the notify
field rather than a control message in order to preserve the ordering
assumptions used by the rest of the protocol. Because the reply is necessary
to order to free both the initiator and target buffers, in contrast to the
existing code, each request requires an explicit reply or empty credit.

Credits are now managed by the buffer allocation machinery rather than being
a separate count.

Subsequent gpd transmission in the local handler is now done using chained
gpds instead of overwriting the first.