Small cleanup w.r.t. "control messages".
Since AM credits no longer use this path, we can predict-false since
the only remaining consumer is shutdown coordination.  Additionally,
there is no need for the switch when there is only one valid "op".

It is possible, but not certain, that we'll remove this path entirely
in a later commit.