Reimplement polling based on peer-list / restore tracing of AM traffic
This commit replaces a array (bounded) of peer numbers with a single-linked
list (unbounded) for the purpose of tracking which peers have generated
Cq events for which we've not yet located a corresponding notify word.
This implementation leads to hold the ampoll lock for a shorter time,
and prevents a small number of peers with highly a degree of re-ordered
packets from "saturating" the list (and thus preventing reception of AMs
from other peers until the re-ordered messages have drained).

This commit also restores and updates the "D" (detail) tracing of the
AM traffic.  It had been disabled in the commit that brought in the
"notify word" support.