Eliminate CQWRITE in favor of a new "type" in the notify ring
The critical path length in Poll should be essentially the same, since either
implementation requires exactly one branch per event (at event_data decode
for CQWRITE, or at notify read time for the notify ring).  In theory the
addition of an additional event in the notify ring could be an added source
of delay due to out-of-order arrivals.  However, currently the only type
of control message sent via this path is the shutdown coordination messages
which should normally be the last things event to ever arrive.

The net win is in code size for not having a distinct send path for the
control msgs.  Timings are indistingishable.