Expose memory registration failures up one level
This commit makes minor changes to the control logic in the bulk Put and
Get paths to move the fatal error handling for bug3178 up one level in
the code.  Later commits will either handle the recovery at this level,
or will pass the failure up one more level (the choice is still TBD).