Implement partial recovery from failed MemRegister calls
This commit modifies the control logic in gasnetc_rdma_put_bulk() and
gasnetc_rdma_get() in the event of a failed MemRegister call.  They
now respond to the failure by falling back on a maximum-sized bounce-
buffer operation to issue a PORTION of the callers request.

Additionally, both functions now return the number of bytes transfered,
which allows the caller to detect the partially-completed xfer.  AND,
since the return in such a case is the maximum bounce-buffer size, the
caller will also learn the proper size for re-issuing the remainder of
the operation without triggering dynamic memory registration which is
bound to fail again.  This makes the very reasonable assumption that
the remainder of the xfer is in the same (remote) segment, but is still
correct (but sub-optimal) regardless.

Logic in the Core (for Long payloads) and Extended will come soon.