Fix bug3178 with respect to Extended API (not yet Core)
This commit revises the main Put and Get "chunking loops" in
gasnet_extended.c to observe the return values from the lower-level
xfer functions.  This allows them to deal *correctly* with the partial
xfers which result when MemRegister fails, by only advancing their
counts by the value returned.  It also allows them to behavior more
nearly optimally in this case, by setting their chunk size to match
the return from the previous xfer (thus going down the bounce-buffer
path for the remainder of the xfer, without the overhead of a
failed MemRegister call).  The actual complexity of these main
Put and Get functions is actually reduced a bit by this change.

Note: movement of AMLong payloads is still vulnerable to bug 3178