Fix bug 3178 with respect to AMLong (but not yet LongAsync)
As with the fix for the Extended API, the logic to Put the Long
payload now uses the return value from gasnetc_rdma_put_bulk() to
issue additional Puts if necessary to deal with a request which
was only partially completed due to a MemRegister failure.

LongAsync will require additional machinery.